moves files into subcategory folder
