+++ futurehope-ReactNative/components/notetaking/NoteTaking.js
@@ -3,7 +3,7 @@ import { View, Text, StyleSheet, TouchableOpacity, FlatList, AsyncStorage, Modal
 import ListItem from './ListItem';
 import AddItem from './AddItem';
 import Icon from "react-native-vector-icons/Entypo"
-import { widthPercentageToDP, heightPercentageToDP } from "../utils/PercenatageFix";
+import { widthPercentageToDP, heightPercentageToDP } from "../../utils/PercenatageFix";
 import { NativeRouter, Route, Link } from "react-router-native";
